Trying to configure login form, deleting sign up file and adding method to UserRepository
