+++ src/Repository/UserRepository.php
@@ -5,6 +5,7 @@ namespace App\Repository;
 
 use App\Model\DatabaseConnection;
 use App\Model\User;
+use PDO;
 
 class UserRepository
 {
@@ -23,5 +24,20 @@ class UserRepository
         $user->mail = $row['mail'];
 
         return $user;
+    }
+
+    public function addUser(string $pseudo, string $mail, string $password): bool
+    {
+        $pwd = password_hash($password, PASSWORD_DEFAULT);
+        $statement = $this->connection->getConnection()->prepare(
+            "INSERT INTO problog.user(pseudo, mail, password) VALUES (?, ?, ?)"
+        );
+        $affectedLine = $statement->execute(array(
+            'pseudo' => $pseudo,
+            'mail'=> $mail,
+            'password' => $pwd)
+        );
+
+        return ($affectedLine > 0);
     }
 }
